requirements.txt
Pillow
pandas
scikit-learn